till mailchimp api

+++ routes/web.php
@@ -20,39 +20,42 @@ Route::get('/', function () {
 
 Route::get('/dashboard', function () {
     return view('dashboard');
-})->middleware(['auth'])->name('dashboard');
+})->middleware(['auth','admin'])->name('dashboard');
 
 //route for products view
 Route::get('/products', function () {
     return view('products');
-})->middleware(['auth'])->name('products');
+})->middleware(['auth','admin'])->name('products');
 
 // route for add product view
 Route::get('/addProduct', function () {
     return view('addProduct');
-})->middleware(['auth'])->name('addProducts');
+})->middleware(['auth','admin'])->name('addProducts');
 
 // create a new product 
-Route::get('product',[ProductController::class,'create']);
-Route::post('product',[ProductController::class,'store']);
+Route::get('product',[ProductController::class,'create'])->middleware(['auth','admin'])->name('create');
+Route::post('product',[ProductController::class,'store'])->middleware(['auth','admin'])->name('store');
 
 // show products for admin for crud
-Route::get('/products',[ProductController::class,'show'])->middleware(['auth'])->name('products');
+Route::get('/products',[ProductController::class,'show'])->middleware(['auth','admin'])->name('products');
 
 // edit product view route 
-Route::get('product/edit/{id}',[ProductController::class,'edit'])->middleware(['auth'])->name('productedit');
+Route::get('product/edit/{id}',[ProductController::class,'edit'])->middleware(['auth','admin'])->name('productedit');
 
 // edit product POST or submit data route
-Route::post('editproduct/{id}',[ProductController::class,'update'])->middleware(['auth'])->name('productedit');
+Route::post('editproduct/{id}',[ProductController::class,'update'])->middleware(['auth','admin'])->name('productedit');
 
 // del product
-Route::delete('/products/{products}',[ProductController::class,'destroy'])->middleware(['auth'])->name('productdelete');
+Route::delete('/products/{products}',[ProductController::class,'destroy'])->middleware(['auth','admin'])->name('productdelete');
 
 // show all products to homepage for user
 Route::get('/',[ProductController::class,'showProduct']);
 
-
-
+// mailchimp
+Route::post('newsletter',[ProductController::class,'newsLetter']);
 
 require __DIR__.'/auth.php';
+
+
+
 ?>
